[TASK] Split build artifact in several artifact classes
